requirements.txt
Django==4.2.6
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
